backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
mysql-connector-python==8.3.0
duckdb==0.9.2
clickhouse-driver==0.2.6
requests==2.31.0
python-dotenv==1.0.1
pydantic==2.6.1
psycopg2-binary==2.9.9
duckdb-engine==0.9.2
sqlalchemy==2.0.27
pymysql==1.1.0
cryptography==42.0.2
clickhouse-connect==0.7.0
cloudflare==2.8.15
pyodbc==5.0.1
openai==1.12.0
asyncpg==0.29.0 